ai-api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
